webcam_server/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
opencv-python-headless==4.5.3.56
pyyaml==5.4.1
requests==2.25.1
python-dotenv==0.15.0
requests-oauthlib==1.3.0
gunicorn==20.1.0
numpy==1.21.0
werkzeug==2.0.3
onvif-zeep==0.2.12
flask-login==0.6.2
flask-talisman==1.0.0
pyjwt==2.6.0
marshmallow==3.19.0
cryptography==39.0.1
redis==4.5.4